requirements.txt
langchain
python-dotenv
streamlit
langchain_community
langchain-openai
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu

arxiv
wikipedia
langchainhub

groq
beautifulsoup4
langchain-groq